pyproject.toml
[tool.poetry]
name = "mymodule"
version = "0.1.0"
description = "generic descripction for generic module"
authors = ["Diego Reiriz Cores <dreiriz@gradiant.org>"]

[tool.poetry.scripts]
mymodule="mymodule.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
SQLAlchemy = "^1.4.39"
duckdb = "^0.6.1"
fastapi = "^0.89.1"
pytest = "^7.2.1"
python-dotenv = "^0.21.1"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
pylint = "^2.14.5"
black = "^22.6.0"
flake8 = "^4.0.1"
pre-commit = "^2.20.0"

[tool.poetry.group.test.dependencies]
pytest-cov = "^3.0.0"
pytest = "^7.1.2"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest]
env_override_existing_values = 1
env_files = [
    ".env",
    ".env.test"
]
log_cli = true
log_level = "DEBUG"

